Chore: initial commit
In this commit, we are uploading the current export of the work so far with Guitar Gallery's website. The plan is to use this docker setup as the dev environment. Once the site is ready to go to stage, we can use the export feature here to export the site contents and upload onto a hosted stage site for stakeholders.